+++ src/main/java/UMC/news/newsIntelligent/domain/notification/entity/NotificationType.java
@@ -1,4 +1,4 @@
-package UMC.news.newsIntelligent.domain.notification;
+package UMC.news.newsIntelligent.domain.notification.entity;
 
 public enum NotificationType {
 	SUBSCRIBED,	// 구독한 토픽 알림
